[IMP] coupon_incompatibility, sale_coupon_incompatibility
Allow coupon_incompatibility settings to be used by other modules.
