+++ coupon_incompatibility/README.rst
@@ -14,10 +14,10 @@ Coupon incompatibilities
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github
-    :target: https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_incompatibility
+    :target: https://github.com/OCA/sale-promotion/tree/15.0/coupon_incompatibility
     :alt: OCA/sale-promotion
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-sale_coupon_incompatibility
+    :target: https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-coupon_incompatibility
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
     :target: https://runboat.odoo-community.org/webui/builds.html?repo=OCA/sale-promotion&target_branch=15.0
@@ -25,22 +25,21 @@ Coupon incompatibilities
 
 |badge1| |badge2| |badge3| |badge4| |badge5| 
 
-This module allows to set incompatibilities between promotions, so when a promotion A
-is not compatible with promotion B, they can't be applied together in a sales order.
+Base module to handle incompatibility between promotions and coupons
 
 **Table of contents**
 
 .. contents::
    :local:
 
-Usage
-=====
+Configuration
+=============
 
-Once you've configured your incompatibilities, go to a quotations that fulfills your
-incompatible promotions and try to set them altogether.
+To configure promotion incompatibilities (either coupons or promotion programs):
 
-- A coupon program with no code, won't be applied.
-- A coupon or program code, will raise an error.
+#. Go to *Sales > Products > Coupon Programs* and select or create a new one.
+#. In the *Incompatible Promotions* set the promotions that aren't compatible with the
+   one you're configuring.
 
 Bug Tracker
 ===========
@@ -48,7 +47,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/sale-promotion/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/sale-promotion/issues/new?body=module:%20sale_coupon_incompatibility%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/sale-promotion/issues/new?body=module:%20coupon_incompatibility%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -90,6 +89,6 @@ Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
 
 |maintainer-chienandalu| 
 
-This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_incompatibility>`_ project on GitHub.
+This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/15.0/coupon_incompatibility>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
