+++ coupon_incompatibility/i18n/coupon_incompatibility.pot
@@ -1,0 +1,26 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* coupon_incompatibility
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 15.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2023-02-02 10:25+0000\n"
+"PO-Revision-Date: 2023-02-02 10:25+0000\n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: coupon_incompatibility
+#: model:ir.model,name:coupon_incompatibility.model_coupon_program
+msgid "Coupon Program"
+msgstr ""
+
+#. module: coupon_incompatibility
+#: model:ir.model.fields,field_description:coupon_incompatibility.field_coupon_program__incompatible_promotion_ids
+msgid "Incompatible Promotions"
+msgstr ""
