+++ coupon_incompatibility/models/__init__.py
@@ -1,2 +1,1 @@
 from . import coupon_program
-from . import sale_coupon
