+++ coupon_incompatibility/readme/CONFIGURE.rst
@@ -1,0 +1,5 @@
+To configure promotion incompatibilities (either coupons or promotion programs):
+
+#. Go to *Sales > Products > Coupon Programs* and select or create a new one.
+#. In the *Incompatible Promotions* set the promotions that aren't compatible with the
+   one you're configuring.
