+++ coupon_incompatibility/readme/CONTRIBUTORS.rst
@@ -1,0 +1,5 @@
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+  * Pedro M. Baeza
+  * David Vidal
+  * Stefan Ungureanu
