+++ coupon_incompatibility/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Base module to handle incompatibility between promotions and coupons
