+++ coupon_incompatibility/tests/test_coupon_incompatibility.py
@@ -1,0 +1,36 @@
+# Copyright 2021 Tecnativa - David Vidal
+# Copyright 2023 Tecnativa - Stefan Ungureanu
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+from odoo.tests import TransactionCase
+
+
+class CouponIncompatibilityCase(TransactionCase):
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        product_obj = cls.env["product.product"]
+        cls.pricelist = cls.env["product.pricelist"].create(
+            {
+                "name": "Test pricelist",
+                "item_ids": [
+                    (
+                        0,
+                        0,
+                        {
+                            "applied_on": "3_global",
+                            "compute_price": "formula",
+                            "base": "list_price",
+                        },
+                    )
+                ],
+            }
+        )
+        cls.partner_1 = cls.env["res.partner"].create(
+            {"name": "Mr. Odoo", "property_product_pricelist": cls.pricelist.id}
+        )
+        cls.partner_2 = cls.env["res.partner"].create(
+            {"name": "Mrs. Odoo", "property_product_pricelist": cls.pricelist.id}
+        )
+        cls.product_a = product_obj.create({"name": "Product A"})
+        cls.product_b = product_obj.create({"name": "Product B"})
+        cls.product_c = product_obj.create({"name": "Product C"})
