+++ sale_coupon_incompatibility/i18n/es.po
@@ -7,23 +7,25 @@ msgstr ""
 "Project-Id-Version: Odoo Server 15.0\n"
 "Report-Msgid-Bugs-To: \n"
 "POT-Creation-Date: 2023-02-02 10:28+0000\n"
-"PO-Revision-Date: 2023-02-02 10:28+0000\n"
+"PO-Revision-Date: 2023-02-02 11:30+0100\n"
 "Last-Translator: \n"
 "Language-Team: \n"
+"Language: es_ES\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+"X-Generator: Poedit 3.2.2\n"
 
 #. module: sale_coupon_incompatibility
 #: model:ir.model,name:sale_coupon_incompatibility.model_coupon_coupon
 msgid "Coupon"
-msgstr ""
+msgstr "Cupón"
 
 #. module: sale_coupon_incompatibility
 #: model:ir.model,name:sale_coupon_incompatibility.model_coupon_program
 msgid "Coupon Program"
-msgstr ""
+msgstr "Programa de cupones"
 
 #. module: sale_coupon_incompatibility
 #: code:addons/sale_coupon_incompatibility/models/sale_coupon.py:0
@@ -33,3 +35,5 @@ msgid ""
 "This promotion is incompatible with other set already in the order so it "
 "can't be applied."
 msgstr ""
+"Esta promoción es incompatible con otra ya presente en la orden por lo "
+"que no se puede aplicar."
