+++ sale_coupon_incompatibility/models/coupon_program.py
@@ -3,14 +3,13 @@
 from odoo import _, models
 
 
-class Coupon(models.Model):
-    _inherit = "coupon.coupon"
+class CouponProgram(models.Model):
+    _inherit = "coupon.program"
 
-    def _check_coupon_code(self, order_date, partner_id, **kwargs):
-        """Coupon incompatibility rules. Check the error strings for a detailed case
+    def _check_promo_code(self, order, coupon_code):
+        """Promotion incompatibility rules. Check the error strings for a detailed case
         detail."""
-        message = super()._check_coupon_code(order_date, partner_id, **kwargs)
-        order = kwargs.get("order", False)
+        message = super()._check_promo_code(order, coupon_code)
         # Other errors may precede
         if message:
             return message
@@ -18,9 +17,7 @@ class Coupon(models.Model):
         order_programs |= (
             order.applied_coupon_ids + order.generated_coupon_ids
         ).mapped("program_id")
-        if any(
-            {x in order_programs for x in self.program_id.incompatible_promotion_ids}
-        ):
+        if any({x in order_programs for x in self.incompatible_promotion_ids}):
             message = {
                 "error": _(
                     "This promotion is incompatible with other set already in the "
