--- sale_coupon_incompatibility/readme/CONFIGURE.rst
@@ -1,7 +1,0 @@
-To configure promotion incompatibilities (either coupons or promotion programs):
-
-#. Go to *Sales > Catalog > Coupon Programs* and select or create a new one.
-#. In the *Incompatible Promotions* set the promotions that aren't compatible with the
-   one you're configuring.
-#. Once set, the those promotions will have your configured promotion as incompatible
-   as well, so they can't be used altogether in the same sale order.
