+++ setup/coupon_incompatibility/odoo/addons/coupon_incompatibility
@@ -1,0 +1,1 @@
+../../../../coupon_incompatibility
